+++ fetchOptionsGraphQL.test.mjs
@@ -16,7 +16,7 @@ export default (tests) => {
   tests.add("`fetchOptionsGraphQL` bundle size.", async () => {
     await assertBundleSize(
       new URL("./fetchOptionsGraphQL.mjs", import.meta.url),
-      1200
+      800
     );
   });